[base] Restructure user-color files
