+++ packages/@sanity/base/src/hooks.ts
@@ -1,4 +1,4 @@
 export {useDocumentPresence, useGlobalPresence} from './datastores/presence/hooks'
 export {useUser} from './datastores/user/hooks'
-export {useUserColor} from './user-color/useUserColor'
+export {useUserColor} from './user-color/hooks'
 export {useTimeAgo} from './time/useTimeAgo'
