+++ packages/@sanity/base/src/user-color/context.ts
@@ -1,0 +1,5 @@
+import {createContext} from 'react'
+import {userColorManager} from './singleton'
+import {UserColorManager} from './types'
+
+export const UserColorManagerContext = createContext<UserColorManager>(userColorManager)
