+++ packages/@sanity/base/src/user-color/index.ts
@@ -1,5 +1,5 @@
-export {createUserColorManager} from './manager'
-export {useUserColorManager, UserColorManagerProvider} from './provider'
-export {userColorManager} from './singleton'
-export {useUserColor} from './useUserColor'
+export * from './hooks'
+export * from './manager'
+export * from './provider'
+export * from './singleton'
 export * from './types'
